pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "u2net_fast"
version = "1.0.0"
authors = [
  { name="Steven Bertolucci", email="srbertol@mtu.edu" },
]
description = "A re-implementation of U2Net to provide better performance with large images."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies=[
    "torch",
    "torchvision",
    "tqdm",
    "pillow"
]

[tool.hatch.build.targets.sdist]
exclude = [
  "examples",
  ".idea",
]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"